+++ packages/extension-koni-ui/src/components/index.ts
@@ -23,6 +23,7 @@ export * from '../contexts';
 export * from './Account';
 export * from './Avatar';
 export * from './Confirmation';
+export * from './EmptyList';
 export * from './Field';
 export * from './History';
 export * from './Icon';